Code cleaning for new design
